requirements.txt
aioitertools==0.12.0
annotated-types==0.7.0
ase==3.25.0
attrs==25.3.0
bcrypt==4.3.0
bibtexparser==1.4.3
boto3==1.38.32
botocore==1.38.32
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
dnspython==2.7.0
emmet-core==0.84.7rc4
fonttools==4.58.2
idna==3.10
jmespath==1.0.1
joblib==1.5.1
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
latexcodec==3.0.0
load-dotenv==0.1.0
maggma==0.71.5
matplotlib==3.10.3
mongomock==4.3.0
monty==2025.3.3
mp-api==0.45.5
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.41.1
networkx==3.5
numpy==2.3.0
orjson==3.10.18
packaging==25.0
palettable==3.3.3
pandas==2.3.0
pandas-stubs==2.2.3.250527
paramiko==3.5.1
pillow==11.2.1
plotly==6.1.2
pybtex==0.24.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydash==8.0.5
pymatgen==2025.5.28
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
ruamel.yaml==0.18.13
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
scipy==1.15.3
sentinels==1.0.0
six==1.17.0
smart-open==7.1.0
spglib==2.6.0
sshtunnel==0.4.0
sympy==1.14.0
tabulate==0.9.0
tqdm==4.67.1
types-pytz==2025.2.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uncertainties==3.2.3
urllib3==2.4.0
wrapt==1.17.2
